requirements.txt
asgiref==3.9.1
atomicwrites==1.4.1
attrs==25.3.0
beautifulsoup4==4.13.4
black==25.1.0
bs4==0.0.2
certifi==2025.8.3
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
crispy-bootstrap4==2025.6
crispy-bootstrap5==2025.6
distlib==0.4.0
Django==5.2.4
django-crispy-forms==2.4
django-debug-toolbar==6.0.0
django-extensions==4.1
filelock==3.18.0
flake8==5.0.4
flake8-annotations==2.9.1
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
gunicorn==23.0.0
h11==0.16.0
identify==2.6.13
idna==3.10
iniconfig==2.1.0
jsonpatch==1.33
jsonpointer==3.0.0
mccabe==0.7.0
mypy_extensions==1.1.0
networkx==3.5
nodeenv==1.9.1
numpy==2.3.2
packaging==25.0
pathspec==0.12.1
pep8-naming==0.13.2
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
plus==1994.12.4
pre_commit==4.3.0
psycopg2-binary==2.9.10
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
pytest-mock==3.8.2
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
scipy==1.16.1
shell==1.0.1
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
toml==0.10.2
tomli==2.2.1
tornado==6.5.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.33.1
visdom==0.2.4
websocket-client==1.8.0
